pytorch

+++ tests/beignet/_polynomial/test_poly2cheb.py
@@ -7,6 +7,8 @@ from .test_polynomial import chebyshev_polynomial_coefficients
 def test_poly2cheb():
     for index in range(10):
         torch.testing.assert_close(
-            beignet.polynomial.poly2cheb(chebyshev_polynomial_coefficients[index]),
-            torch.tensor([0] * index + [1]),
+            beignet.polynomial.poly2cheb(
+                chebyshev_polynomial_coefficients[index],
+            ),
+            torch.tensor([0] * index + [1], dtype=torch.float64),
         )
